+++ brion/spikeReportPlugin.h
@@ -70,7 +70,7 @@ typedef PluginInitData SpikeReportInitData;
  *
  * Plugin libraries in the LD_LIBRARY_PATH will be automatically registered if
  * they provide the abovementioned C functions and follow the naming convention:
- * <lib_prefix>Brion<MyReportName>SpikeReport.<lib_extension>
+ * \<lib_prefix\>Brion\<MyReportName\>SpikeReport.\<lib_extension\>
  *
  * @version 1.4
  */